Fix return typing
File size is always an integer

+++ lib/Service/DropboxStorageAPIService.php
@@ -238,7 +238,7 @@ class DropboxStorageAPIService {
 	 * @return ?int downloaded size, null if already existing or network error
 	 */
 	private function getFile(string $accessToken, string $refreshToken, string $clientID, string $clientSecret,
-							string $userId, array $fileItem, Node $topFolder): ?float {
+							string $userId, array $fileItem, Node $topFolder): ?int {
 		$fileName = $fileItem['name'];
 		$path = preg_replace('/^\//', '', $fileItem['path_display'] ?? '.');
 		$pathParts = pathinfo($path);
